fixed header and footer editors
