+++ frontend/web/churchlink/src/features/admin/components/WebBuilder/sub_pages/EditFooter.tsx
@@ -29,17 +29,6 @@ import {
     DialogTitle,
     DialogTrigger,
 } from "@/shared/components/ui/Dialog";
-import { Card, CardContent, CardHeader, CardTitle } from "@/shared/components/ui/card";
-import { Input } from "@/shared/components/ui/input";
-import {
-    Dialog,
-    DialogContent,
-    DialogDescription,
-    DialogFooter,
-    DialogHeader,
-    DialogTitle,
-    DialogTrigger,
-} from "@/shared/components/ui/Dialog";
 import {
     AlertDialog,
     AlertDialogAction,
@@ -50,7 +39,6 @@ import {
     AlertDialogHeader,
     AlertDialogTitle,
 } from "@/shared/components/ui/alert-dialog";
-import MultiStateBadge from "@/shared/components/MultiStageBadge";
 import MultiStateBadge from "@/shared/components/MultiStageBadge";
 
 interface FooterItem {
@@ -65,10 +53,6 @@ interface FooterSection {
     russian_title: string;
     items: FooterItem[];
     visible?: boolean;
-}
-
-interface LinkItem extends FooterItem {
-    type?: "link";
 }
 
 interface LinkItem extends FooterItem {
@@ -102,69 +86,8 @@ const FooterVisibilityToggle: React.FC<{ section: FooterSection; onToggle: (titl
             setTimeout(() => setBadgeState("custom"), 1200);
         }
     };
-// Visibility toggle component using MultiStateBadge
-const FooterVisibilityToggle: React.FC<{ section: FooterSection; onToggle: (title: string, currentVisibility: boolean) => void }> = ({ section, onToggle }) => {
-    const [badgeState, setBadgeState] = useState<"custom" | "processing" | "success" | "error">("custom");
-
-    const handleToggleVisibility = async () => {
-        if (badgeState !== "custom") return;
-        const currentVisibility = section.visible ?? false;
-        const newVisibility = !currentVisibility;
-        setBadgeState("processing");
-
-        try {
-            await api.put(`/v1/footer/${section.title}/visibility`, { visible: newVisibility });
-            setBadgeState("success");
-            setTimeout(() => {
-                onToggle(section.title, currentVisibility);
-                setBadgeState("custom");
-            }, 900);
-        } catch (error) {
-            console.error("Error updating footer visibility:", error);
-            setBadgeState("error");
-            setTimeout(() => setBadgeState("custom"), 1200);
-        }
-    };
 
     return (
-        <div className="flex items-center justify-center h-full w-full overflow-visible">
-            <MultiStateBadge
-                state={badgeState}
-                onClick={handleToggleVisibility}
-                customComponent={
-                    <span
-                        className={`inline-block px-2 py-1 text-xs rounded-full font-medium cursor-pointer ${
-                            section.visible
-                                ? "bg-green-100 text-green-800"
-                                : "bg-red-100 text-red-800"
-                        }`}
-                    >
-                        {section.visible ? "Visible" : "Hidden"}
-                    </span>
-                }
-            />
-        </div>
-    );
-};
-
-const SortableItem = ({ id, children }: { id: string; children: React.ReactNode }) => {
-    const { attributes, listeners, setNodeRef, transform, transition } = useSortable({ id });
-
-    const style = {
-        transform: CSS.Transform.toString(transform),
-        transition,
-    };
-
-    return (
-        <div ref={setNodeRef} style={style} className="border-b last:border-0">
-            <div className="flex items-center">
-                <div {...attributes} {...listeners} className="cursor-grab p-2 mr-2 text-gray-400">
-                    &#x2630;
-                </div>
-                <div className="flex-grow">
-                    {children}
-                </div>
-            </div>
         <div className="flex items-center justify-center h-full w-full overflow-visible">
             <MultiStateBadge
                 state={badgeState}
@@ -240,32 +163,6 @@ const EditFooter = ({ onFooterDataChange }: EditFooterProps = {}) => {
     const [editTempItemUrl, setEditTempItemUrl] = useState("");
     const [editingItemIndex, setEditingItemIndex] = useState<number | null>(null);
 
-    const [hasUnsavedChanges, setHasUnsavedChanges] = useState(false);
-
-    // Modal states
-    const [isAddModalOpen, setIsAddModalOpen] = useState(false);
-    const [isEditModalOpen, setIsEditModalOpen] = useState(false);
-    const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
-    const [itemToDelete, setItemToDelete] = useState<string | null>(null);
-
-    // Add section form state
-    const [newSectionTitle, setNewSectionTitle] = useState("");
-    const [newSectionRussianTitle, setNewSectionRussianTitle] = useState("");
-    const [sectionItems, setSectionItems] = useState<FooterItem[]>([]);
-    const [tempItemTitle, setTempItemTitle] = useState("");
-    const [tempItemRussianTitle, setTempItemRussianTitle] = useState("");
-    const [tempItemUrl, setTempItemUrl] = useState("");
-
-    // Edit section form state
-    const [editingSection, setEditingSection] = useState<FooterSection | null>(null);
-    const [editSectionTitle, setEditSectionTitle] = useState("");
-    const [editSectionRussianTitle, setEditSectionRussianTitle] = useState("");
-    const [editSectionItems, setEditSectionItems] = useState<LinkItem[]>([]);
-    const [editTempItemTitle, setEditTempItemTitle] = useState("");
-    const [editTempItemRussianTitle, setEditTempItemRussianTitle] = useState("");
-    const [editTempItemUrl, setEditTempItemUrl] = useState("");
-    const [editingItemIndex, setEditingItemIndex] = useState<number | null>(null);
-
     const sensors = useSensors(useSensor(PointerSensor));
 
     useEffect(() => {
@@ -285,9 +182,6 @@ const EditFooter = ({ onFooterDataChange }: EditFooterProps = {}) => {
             const response = await api.get("/v1/footer/items");
             setFooter(response.data);
             setHasUnsavedChanges(false);
-            const response = await api.get("/v1/footer/items");
-            setFooter(response.data);
-            setHasUnsavedChanges(false);
         } catch (err) {
             console.error("Failed to fetch footer:", err);
             toast.error("Failed to load footer data");
@@ -300,199 +194,6 @@ const EditFooter = ({ onFooterDataChange }: EditFooterProps = {}) => {
         if (!footer) return;
 
         const { active, over } = event;
-        if (!over || active.id === over.id) return;
-
-        const oldIndex = footer.items.findIndex((s) => s.title === active.id);
-        const newIndex = footer.items.findIndex((s) => s.title === over.id);
-        if (oldIndex === -1 || newIndex === -1) return;
-
-        const items = arrayMove(footer.items, oldIndex, newIndex);
-        setFooter({ ...footer, items });
-        setDirty(true);
-    };
-
-    const saveAll = async () => {
-        if (!footer) return;
-        try {
-            // Save reordering
-            const currentTitles = footer.items.map(item => item.title);
-            await api.put("/v1/footer/reorder", { titles: currentTitles });
-            toast.success("Footer changes saved successfully");
-
-            // Refresh data from server
-            await fetchFooter();
-        } catch (err: any) {
-            console.error("Failed to save footer changes:", err);
-            toast.error("Failed to save changes");
-            await fetchFooter(); // Revert to server state on failure
-        }
-    };
-
-    // Modal handlers
-    const handleAddSection = async (e: React.FormEvent) => {
-        e.preventDefault();
-        if (!newSectionTitle) {
-            toast.error("Section title is required");
-            return;
-        }
-
-        if (!newSectionRussianTitle) {
-            toast.error("Section russian title is required");
-            return;
-        }
-
-        try {
-            const res = await api.post("/v1/footer/items", {
-                "title": newSectionTitle,
-                "russian_title": newSectionRussianTitle,
-                "items": sectionItems,
-                "visible": false,
-            });
-            if (res) {
-                toast.success("Section added successfully!");
-                // Reset form
-                setNewSectionTitle("");
-                setNewSectionRussianTitle("");
-                setSectionItems([]);
-                setIsAddModalOpen(false);
-                // Refresh footer data
-                await fetchFooter();
-            }
-        } catch (err) {
-            console.error("Failed to add section:", err);
-            toast.error("Failed to add section");
-        }
-    };
-
-    const handleAddSectionItem = (e: React.FormEvent) => {
-        e.preventDefault();
-        if (!tempItemTitle || !tempItemRussianTitle) {
-            toast.error("Item title and Russian title are required");
-            return;
-        }
-        setSectionItems([...sectionItems, { title: tempItemTitle, russian_title: tempItemRussianTitle, url: tempItemUrl || null }]);
-        setTempItemTitle("");
-        setTempItemRussianTitle("");
-        setTempItemUrl("");
-    };
-
-    const handleRemoveSectionItem = (index: number) => {
-        setSectionItems(sectionItems.filter((_, i) => i !== index));
-    };
-
-    const handleEditSection = (section: FooterSection) => {
-        setEditingSection(section);
-        setEditSectionTitle(section.title);
-        setEditSectionRussianTitle(section.russian_title);
-        setEditSectionItems(section.items);
-        setIsEditModalOpen(true);
-    };
-
-    const handleEditSubmit = async (e: React.FormEvent) => {
-        e.preventDefault();
-
-        if (!editingSection) return;
-
-        try {
-            const updatedSection = {
-                title: editSectionTitle,
-                russian_title: editSectionRussianTitle,
-                items: editSectionItems,
-            };
-
-            const response = await api.put(`/v1/footer/items/edit/${editingSection.title}`, updatedSection);
-
-            if (response.data) {
-                toast.success("Section updated successfully");
-                setIsEditModalOpen(false);
-                setEditingSection(null);
-                await fetchFooter();
-            }
-        } catch (err) {
-            console.error("Failed to update section:", err);
-            toast.error("Failed to update section");
-        }
-    };
-
-    const handleEditAddItem = (e: React.FormEvent) => {
-        e.preventDefault();
-
-        if (!editTempItemTitle || !editTempItemRussianTitle) {
-            toast.error("Item title and Russian title are required");
-            return;
-        }
-
-        if (editingItemIndex !== null) {
-            // Update existing item
-            const updatedItems = [...editSectionItems];
-            updatedItems[editingItemIndex] = {
-                title: editTempItemTitle,
-                russian_title: editTempItemRussianTitle,
-                url: editTempItemUrl || null,
-                visible: true,
-                type: "link"
-            };
-            setEditSectionItems(updatedItems);
-            setEditingItemIndex(null);
-        } else {
-            // Add new item
-            setEditSectionItems([...editSectionItems, {
-                title: editTempItemTitle,
-                russian_title: editTempItemRussianTitle,
-                url: editTempItemUrl || null,
-                visible: true,
-                type: "link"
-            }]);
-        }
-
-        // Reset form fields
-        setEditTempItemTitle("");
-        setEditTempItemRussianTitle("");
-        setEditTempItemUrl("");
-    };
-
-    const handleEditItem = (index: number) => {
-        const item = editSectionItems[index];
-        setEditTempItemTitle(item.title);
-        setEditTempItemRussianTitle(item.russian_title);
-        setEditTempItemUrl(item.url || "");
-        setEditingItemIndex(index);
-    };
-
-    const handleRemoveEditItem = (index: number) => {
-        setEditSectionItems(editSectionItems.filter((_, i) => i !== index));
-        // If we were editing this item, reset the form
-        if (editingItemIndex === index) {
-            setEditTempItemTitle("");
-            setEditTempItemRussianTitle("");
-            setEditTempItemUrl("");
-            setEditingItemIndex(null);
-        }
-    };
-
-    const cancelEditingItem = () => {
-        setEditTempItemTitle("");
-        setEditTempItemRussianTitle("");
-        setEditTempItemUrl("");
-        setEditingItemIndex(null);
-    };
-
-    const handleRemoveSection = (title: string) => {
-        setItemToDelete(title);
-        setIsDeleteModalOpen(true);
-    };
-
-    const confirmDeleteSection = async () => {
-        if (!itemToDelete) return;
-
-        try {
-            await api.delete(`/v1/footer/${itemToDelete}`);
-            toast.success(`"${itemToDelete}" removed successfully`);
-
-            if (footer) {
-                // Update UI after successful deletion
-                const newItems = footer.items.filter(item => item.title !== itemToDelete);
-                setFooter({ ...footer, items: newItems });
         if (active.id !== over?.id) {
             const oldIndex = footer.items.findIndex(item => item.title === active.id);
             const newIndex = footer.items.findIndex(item => item.title === over?.id);
@@ -691,30 +392,12 @@ const EditFooter = ({ onFooterDataChange }: EditFooterProps = {}) => {
             }
 
             // Refresh data from server to ensure consistency
-            // Refresh data from server to ensure consistency
             await fetchFooter();
         } catch (err) {
             console.error("Failed to remove section:", err);
             toast.error(`Failed to remove "${itemToDelete}"`);
             // Refresh to revert any optimistic UI updates
-        } catch (err) {
-            console.error("Failed to remove section:", err);
-            toast.error(`Failed to remove "${itemToDelete}"`);
-            // Refresh to revert any optimistic UI updates
             await fetchFooter();
-        } finally {
-            setIsDeleteModalOpen(false);
-            setItemToDelete(null);
-        }
-    };
-
-    const handleChangeVisibility = (title: string, currentVisibility: boolean) => {
-        // Update local state after successful API call from FooterVisibilityToggle
-        if (footer) {
-            const newItems = footer.items.map(item =>
-                item.title === title ? { ...item, visible: !currentVisibility } : item
-            );
-            setFooter({ ...footer, items: newItems });
         } finally {
             setIsDeleteModalOpen(false);
             setItemToDelete(null);
@@ -839,381 +522,6 @@ const EditFooter = ({ onFooterDataChange }: EditFooterProps = {}) => {
                                     </div>
                                 </div>
 
-                                <DialogFooter>
-                                    <Button type="submit" disabled={sectionItems.length === 0}>
-                                        Add Section
-                                    </Button>
-                                </DialogFooter>
-                            </form>
-                        </DialogContent>
-                    </Dialog>
-
-                    {/* Edit Section Dialog */}
-                    <Dialog open={isEditModalOpen} onOpenChange={setIsEditModalOpen}>
-                        <DialogContent className="sm:max-w-[600px]">
-                            <DialogHeader>
-                                <DialogTitle>Edit Footer Section</DialogTitle>
-                                <DialogDescription>
-                                    Update the section details and items.
-                                </DialogDescription>
-                            </DialogHeader>
-                            <form onSubmit={handleEditSubmit} className="flex flex-col gap-4">
-                                <Input
-                                    type="text"
-                                    placeholder="Section Title"
-                                    value={editSectionTitle}
-                                    onChange={(e) => setEditSectionTitle(e.target.value)}
-                                    required
-                                />
-                                <Input
-                                    type="text"
-                                    placeholder="Russian Title"
-                                    value={editSectionRussianTitle}
-                                    onChange={(e) => setEditSectionRussianTitle(e.target.value)}
-                                    required
-                                />
-
-                                <div className="border rounded p-4 bg-gray-50">
-                                    <h4 className="font-medium mb-2">Section Items</h4>
-
-                                    {editSectionItems.length > 0 ? (
-                                        <DndContext
-                                            sensors={sensors}
-                                            collisionDetection={closestCenter}
-                                            onDragEnd={handleDragEnd}
-                                        >
-                                            <SortableContext
-                                                items={editSectionItems.map((link) => `${link.title}-${link.url}`)}
-                                                strategy={verticalListSortingStrategy}
-                                            >
-                                                <ul className="mb-4">
-                                                    {editSectionItems.map((link, index) => (
-                                                        <SortableItem
-                                                            key={`${link.title}-${link.url}`}
-                                                            id={`${link.title}-${link.url}`}
-                                                        >
-                                                            <li className="flex justify-between items-center p-2 bg-white border rounded">
-                                                                <div>
-                                                                    <div className="font-medium">{link.title}</div>
-                                                                    <div className="text-sm text-blue-600">{link.url}</div>
-                                                                </div>
-                                                                <div className="flex gap-2">
-                                                                    <Button
-                                                                        type="button"
-                                                                        variant="ghost"
-                                                                        size="sm"
-                                                                        onClick={() => handleEditItem(index)}
-                                                                        className="text-blue-500 hover:text-blue-700"
-                                                                    >
-                                                                        Edit
-                                                                    </Button>
-                                                                    <Button
-                                                                        type="button"
-                                                                        variant="ghost"
-                                                                        size="sm"
-                                                                        onClick={() => handleRemoveEditItem(index)}
-                                                                        className="text-red-500 hover:text-red-700"
-                                                                    >
-                                                                        Remove
-                                                                    </Button>
-                                                                </div>
-                                                            </li>
-                                                        </SortableItem>
-                                                    ))}
-                                                </ul>
-                                            </SortableContext>
-                                        </DndContext>
-                                    ) : (
-                                        <p className="text-gray-500 italic mb-4">No items</p>
-                                    )}
-
-                                    <div className="border-t pt-3">
-                                        <h5 className="font-medium mb-2">
-                                            {editingItemIndex !== null ? 'Edit Item' : 'Add Item to Section'}
-                                        </h5>
-                                        <div className="flex flex-col gap-2">
-                                            <Input
-                                                type="text"
-                                                placeholder="Title"
-                                                value={editTempItemTitle}
-                                                onChange={(e) => setEditTempItemTitle(e.target.value)}
-                                            />
-                                            <Input
-                                                type="text"
-                                                placeholder="Russian Title"
-                                                value={editTempItemRussianTitle}
-                                                onChange={(e) => setEditTempItemRussianTitle(e.target.value)}
-                                            />
-                                            <Input
-                                                type="text"
-                                                placeholder="Optional URL"
-                                                value={editTempItemUrl}
-                                                onChange={(e) => setEditTempItemUrl(e.target.value)}
-                                            />
-                                            <div className="flex gap-2">
-                                                <Button
-                                                    type="button"
-                                                    onClick={handleEditAddItem}
-                                                    size="sm"
-                                                    className="self-start"
-                                                >
-                                                    {editingItemIndex !== null ? 'Update Item' : 'Add Item'}
-                                                </Button>
-                                                {editingItemIndex !== null && (
-                                                    <Button
-                                                        type="button"
-                                                        onClick={cancelEditingItem}
-                                                        size="sm"
-                                                        variant="outline"
-                                                    >
-                                                        Cancel
-                                                    </Button>
-                                                )}
-                                            </div>
-                                        </div>
-                                    </div>
-                                </div>
-
-                                <DialogFooter>
-                                    <Button type="submit">
-                                        Save Changes
-                                    </Button>
-                                </DialogFooter>
-                            </form>
-                        </DialogContent>
-                    </Dialog>
-
-                    {/* Delete Confirmation Dialog */}
-                    <AlertDialog open={isDeleteModalOpen} onOpenChange={setIsDeleteModalOpen}>
-                        <AlertDialogContent>
-                            <AlertDialogHeader>
-                                <AlertDialogTitle>Are you sure?</AlertDialogTitle>
-                                <AlertDialogDescription>
-                                    This action cannot be undone. This will permanently remove the footer section "{itemToDelete}" and all its items.
-                                </AlertDialogDescription>
-                            </AlertDialogHeader>
-                            <AlertDialogFooter>
-                                <AlertDialogCancel>Cancel</AlertDialogCancel>
-                                <AlertDialogAction onClick={confirmDeleteSection} className="bg-red-600 hover:bg-red-700">
-                                    Delete
-                                </AlertDialogAction>
-                            </AlertDialogFooter>
-                        </AlertDialogContent>
-                    </AlertDialog>
-                </div>
-            </CardHeader>
-
-            <CardContent>
-
-            <Card>
-                <CardHeader className="py-4">
-                    <CardTitle className="text-base">Current Sections</CardTitle>
-                    <CardDescription className="text-xs">The list below reflects your pending edits.</CardDescription>
-                </CardHeader>
-
-                <CardContent className="space-y-1.5">
-                    {footer && footer.items.length > 0 ? (
-                        <DndContext sensors={sensors} collisionDetection={closestCenter} onDragEnd={handleDragEnd}>
-                            <SortableContext items={footer.items.map((s) => s.title)} strategy={verticalListSortingStrategy}>
-                                <div className="space-y-1.5">
-                                    {footer.items.map((section) => {
-                                        const visible = !!effectiveVisible(section);
-                                        return (
-                                            <SortableItem key={section.title} id={section.title}>
-                                                <div className="flex w-full items-center justify-between gap-2">
-                                                    <div className="min-w-0">
-                                                        <div className="truncate text-sm font-medium">{section.title}</div>
-                                                        <div className="mt-0.5 flex items-center gap-1.5 text-[11px] text-muted-foreground">
-                                                            <Badge variant="secondary" className="px-1 py-0 text-[10px]">
-                                                                {section.items.length} item{section.items.length === 1 ? "" : "s"}
-                                                            </Badge>
-                                                        </div>
-                                                    </div>
-        <Card className="w-full max-w-4xl mx-auto">
-            <CardHeader>
-                <div className="flex justify-between items-center">
-                    <CardTitle>Edit Footer Sections</CardTitle>
-                    <Dialog open={isAddModalOpen} onOpenChange={setIsAddModalOpen}>
-                        <DialogTrigger asChild>
-                            <Button variant="default">
-                                Add Section
-                            </Button>
-                        </DialogTrigger>
-                        <DialogContent className="sm:max-w-[600px]">
-                            <DialogHeader>
-                                <DialogTitle>Add Footer Section</DialogTitle>
-                                <DialogDescription>
-                                    Create a new footer section with items.
-                                </DialogDescription>
-                            </DialogHeader>
-                            <form onSubmit={handleAddSection} className="flex flex-col gap-4">
-                                <Input
-                                    type="text"
-                                    placeholder="Section Title"
-                                    value={newSectionTitle}
-                                    onChange={(e) => setNewSectionTitle(e.target.value)}
-                                    required
-                                />
-                                <Input
-                                    type="text"
-                                    placeholder="Russian Title"
-                                    value={newSectionRussianTitle}
-                                    onChange={(e) => setNewSectionRussianTitle(e.target.value)}
-                                    required
-                                />
-                                <div className="border rounded p-4 bg-gray-50">
-                                    <h4 className="font-medium mb-2">Section Items</h4>
-
-                                    {sectionItems.length > 0 ? (
-                                        <DndContext
-                                            sensors={sensors}
-                                            collisionDetection={closestCenter}
-                                            onDragEnd={handleDragEnd}
-                                        >
-                                            <SortableContext
-                                                items={sectionItems.map((item) => `${item.title}-${item.url}`)}
-                                                strategy={verticalListSortingStrategy}
-                                            >
-                                                <ul className="mb-4">
-                                                    {sectionItems.map((item, index) => (
-                                                        <SortableItem key={`${item.title}-${item.url}`} id={`${item.title}-${item.url}`}>
-                                                            <li className="flex justify-between items-center p-2 bg-white border rounded">
-                                                                <div>
-                                                                    <div className="font-medium">{item.title}</div>
-                                                                    <div className="text-sm text-blue-600">{item.url}</div>
-                                                                </div>
-                                                                <Button
-                                                                    type="button"
-                                                                    variant="ghost"
-                                                                    size="sm"
-                                                                    onClick={() => handleRemoveSectionItem(index)}
-                                                                    className="text-red-500 hover:text-red-700"
-                                                                >
-                                                                    Remove
-                                                                </Button>
-                                                            </li>
-                                                        </SortableItem>
-                                                    ))}
-                                                </ul>
-                                            </SortableContext>
-                                        </DndContext>
-                                    ) : (
-                                        <p className="text-gray-500 italic mb-4">No links</p>
-                                    )}
-
-                                    <div className="border-t pt-3">
-                                        <h5 className="font-medium mb-2">Add Item to Section</h5>
-                                        <div className="flex flex-col gap-2">
-                                            <Input
-                                                type="text"
-                                                placeholder="Title"
-                                                value={tempItemTitle}
-                                                onChange={(e) => setTempItemTitle(e.target.value)}
-                                            />
-                                            <Input
-                                                type="text"
-                                                placeholder="Russian Title"
-                                                value={tempItemRussianTitle}
-                                                onChange={(e) => setTempItemRussianTitle(e.target.value)}
-                                            />
-                                            <Input
-                                                type="text"
-                                                placeholder="Item URL"
-                                                value={tempItemUrl}
-                                                onChange={(e) => setTempItemUrl(e.target.value)}
-                                            />
-                                            <Button
-                                                type="button"
-                                                onClick={handleAddSectionItem}
-                                                size="sm"
-                                                className="self-start"
-                                            >
-                                                Add to Section
-                                            </Button>
-                                        </div>
-                                    </div>
-                                </div>
-
-                                                    <div className="flex items-center gap-0.5">
-                                                        <div className="mr-0.5 flex items-center gap-1">
-                                                            <Switch
-                                                                checked={visible}
-                                                                onCheckedChange={() => toggleVisibility(section.title, visible)}
-                                                                aria-label="Toggle visibility"
-                                                            />
-                                                            {visible ? (
-                                                                <Eye className="h-3.5 w-3.5 text-muted-foreground" />
-                                                            ) : (
-                                                                <EyeOff className="h-3.5 w-3.5 text-muted-foreground" />
-                                                            )}
-                                                        </div>
-
-                                                        <Button
-                                                            variant="ghost"
-                                                            size="icon"
-                                                            className="h-8 w-8"
-                                                            onClick={() =>
-                                                                navigate(`/admin/webbuilder/footer/edit/${encodeURIComponent(section.title)}`)
-                                                            }
-                                                            aria-label="Edit"
-                                                        >
-                                                            <Pencil className="h-4 w-4" />
-                                                        </Button>
-
-                                                        <Button
-                                                            variant="ghost"
-                                                            size="icon"
-                                                            className="h-8 w-8 text-destructive"
-                                                            aria-label="Remove"
-                                                            onClick={() => setToRemove(section.title)}
-                                                        >
-                                                            <Trash2 className="h-4 w-4" />
-                                                        </Button>
-                                                    </div>
-                                                </div>
-                                            </div>
-                                            <div className="flex gap-2">
-                                                <FooterVisibilityToggle section={item} onToggle={handleChangeVisibility} />
-                                                <Button
-                                                    variant="ghost"
-                                                    size="sm"
-                                                    onClick={() => handleEditSection(item)}
-                                                    className="text-blue-600 hover:text-blue-700"
-                                                >
-                                                    Edit
-                                                </Button>
-                                                <Button
-                                                    variant="ghost"
-                                                    size="sm"
-                                                    onClick={() => handleRemoveSection(item.title)}
-                                                    className="text-red-600 hover:text-red-700"
-                                                >
-                                                    Remove
-                                                </Button>
-                                            </div>
-                                        </li>
-                                    </SortableItem>
-                                ))}
-                            </ul>
-                        </SortableContext>
-                    </DndContext>
-                ) : (
-                    <p className="text-gray-500">No sections yet. Click "Add Section" to create one.</p>
-                )}
-            </div>
-
-            {footer && footer.items.length > 0 && (
-                <div className="flex gap-4 justify-start mt-6">
-                    <Button
-                        onClick={handleSaveChanges}
-                        disabled={!hasUnsavedChanges}
-                    >
-                        Save Changes
-                    </Button>
-                </div>
-            )}
-        </CardContent>
-    </Card>
                                 <DialogFooter>
                                     <Button type="submit" disabled={sectionItems.length === 0}>
                                         Add Section
